Refactor to always update a parquet file and save it via class method.
