+++ roman_photoz/create_simulated_catalog.py
@@ -56,7 +56,7 @@ class SimulatedCatalog:
         Placeholder for simulated data.
     """
 
-    def __init__(self, nobj: int = 1000, mag_noise: float = 0.1):
+    def __init__(self, nobj: int = 1000, mag_noise: float = 0.1, seed=None):
         """
         Initializes the SimulatedCatalog class.
         """
@@ -73,6 +73,7 @@ class SimulatedCatalog:
         self._roman_catalog_template = self._read_roman_template_catalog()
         self.filter_list = get_roman_filter_list()
         self.mag_noise = mag_noise
+        self.seed = seed
 
     def _create_column_names(self):
         colnames_list = []
@@ -196,6 +197,7 @@ class SimulatedCatalog:
         output_filename: str = DEFAULT_OUTPUT_CATALOG_FILENAME,
         output_path: str = "",
         return_catalog: bool = False,
+        seed=None,
     ):
         """
         Create a simulated input catalog from the simulated data.
@@ -456,7 +458,11 @@ class SimulatedCatalog:
                     f"Requested {num_lines} lines, but only {total_lines} lines are available."
                 )
 
-            rng = np.random.default_rng()
+            rng = (
+                np.random.default_rng()
+                if self.seed is None
+                else np.random.default_rng(seed=self.seed)
+            )
             random_indices = rng.choice(total_lines, num_lines, replace=False)
             return self.simulated_data[random_indices]
         else:
@@ -530,12 +536,22 @@ def main():
             action="store_true",
             help="Refresh the lib_mag folder by regenerating the synthetic magnitudes.",
         )
+        parser.add_argument(
+            "--seed",
+            type=int,
+            default=None,
+            help="Seed for the random number generator used in picking random lines and adding noise (default: None).",
+        )
         return parser.parse_args()
 
     args = parse_args()
 
     logger.info("Starting simulated catalog creation...")
-    simulated_catalog = SimulatedCatalog(nobj=args.nobj, mag_noise=args.mag_noise)
+    simulated_catalog = SimulatedCatalog(
+        nobj=args.nobj,
+        mag_noise=args.mag_noise,
+        seed=args.seed,
+    )
     simulated_catalog.process(
         output_path=args.output_path,
         output_filename=args.output_filename,
