+++ roman_photoz/regtest/test_integration.py
@@ -27,7 +27,7 @@ def roman_catalog_process():
 
 def test_roman_photoz(tmp_path, dms_logger):
     # create catalog
-    sc = create_simulated_catalog.SimulatedCatalog(nobj=100, mag_noise=0.02)
+    sc = create_simulated_catalog.SimulatedCatalog(nobj=100, mag_noise=0.01, seed=42)
     sc.process(tmp_path, "cat.parquet")
 
     # create instance
@@ -60,7 +60,7 @@ def test_roman_photoz(tmp_path, dms_logger):
         assert np.any(output[col] != 0)
 
     # make sure the redshifts aren't crazy.
-    err_upper_limit = 0.1
+    err_upper_limit = 0.2
     err = mad_std(output["photoz"] - output["redshift_true"])
     dms_logger.info(
         f"""DMS398 MSG: roman-photoz successfully produced output file
