+++ roman_photoz/roman_catalog_process.py
@@ -80,6 +80,10 @@ class RomanCatalogProcess:
         self.estimated = None
         self.default_roman_output_keys = read_output_keys(DEFAULT_OUTPUT_KEYWORDS)
 
+        self.input_filename = None
+        self.output_filename = None
+        self.output_format = None
+
     def _set_config_file(self, config_filename: Union[dict, str] = ""):
         """
         Set the configuration file.
@@ -235,41 +239,49 @@ class RomanCatalogProcess:
 
     def _save_results(
         self,
-        output_filename: Optional[str] = None,
-        output_format: str = "parquet",
     ):
         """
         Save the results to the specified output file.
-
-        Parameters
-        ----------
-        output_filename : str, optional
-            Name of the output file.
-        output_format : str, optional
-            Format to save the results.
-            Supported formats are "parquet" (default) and "asdf".
 
         Raises
         ------
         ValueError
             If there are no results to save.
         """
-        if self.estimated is not None:
-            ancil_data = self.estimated.data.ancil
+
+        if self.output_filename is None:
+            logger.info("No output filename provided. Updating input file in place.")
+            self._update_input(self.input_filename, save_results=True)
+        elif self.output_filename is not None and self.output_format.lower() in [
+            "parquet",
+            "asdf",
+        ]:
+            if self.estimated is not None:
+                self._update_input(
+                    self.input_filename, save_results=False
+                )
+            else:
+                logger.error("No results to save")
+                raise ValueError("No results to save.")
+
+            if self.output_format.lower() == "parquet":
+                import pyarrow.parquet as pq
+
+                pq.write_table(self.results, self.output_filename)
+            elif self.output_format.lower() == "asdf":
+                tree = {"roman_photoz_results": self.results.to_pydict()}
+                with AsdfFile(tree) as af:
+                    af.write_to(self.output_filename)
         else:
-            logger.error("No results to save")
-            raise ValueError("No results to save.")
+            logger.error(
+                f"Unsupported output format: {self.output_format}. Supported formats are 'parquet' and 'asdf'."
+            )
+            raise ValueError(
+                f"Unsupported output format: {self.output_format}. Supported formats are 'parquet' and 'asdf'."
+            )
+        logger.info(f"Results saved to {self.output_filename}.")
 
-        if output_format.lower() == "parquet":
-            ancil_data = Table(ancil_data)
-            ancil_data.write(output_filename, format="parquet", overwrite=True)
-        elif output_format.lower() == "asdf":
-            tree = {"roman_photoz_results": ancil_data}
-            with AsdfFile(tree) as af:
-                af.write_to(output_filename)
-        logger.info(f"Results saved to {output_filename}.")
-
-    def _update_input(self, input_filename):
+    def _update_input(self, input_filename, save_results=False):
         # TODO: this can be done with the Table class
         # directly; no need for pyarrow
         import pyarrow as pa
@@ -331,8 +343,9 @@ class RomanCatalogProcess:
 
         extra_astropy_metadata = astropy.table.meta.get_yaml_from_table(tab_astro)
         meta[b"table_meta_yaml"] = "\n".join(extra_astropy_metadata).encode("utf-8")
-        tab = tab.replace_schema_metadata(meta)
-        pq.write_table(tab, input_filename)
+        self.results = tab.replace_schema_metadata(meta)
+        if save_results:
+            pq.write_table(self.results, self.input_filename)
 
     def process(
         self,
@@ -359,8 +372,14 @@ class RomanCatalogProcess:
             The type of flux to use for fitting.
             Options are "psf" (default), "kron", "segment", or "aperture."
         """
+
+        self.input_filename = input_filename
+        if output_filename is not None:
+            self.output_filename = output_filename
+            self.output_format = output_format
+
         self.data = self._get_data(
-            input_filename=input_filename,
+            input_filename=self.input_filename,
             fit_colname=fit_colname,
             fit_err_colname=fit_err_colname,
         )
@@ -372,13 +391,7 @@ class RomanCatalogProcess:
             self._create_informer_stage()
         self._create_estimator_stage()
 
-        if output_filename is not None:
-            self._save_results(
-                output_filename=output_filename,
-                output_format=output_format,
-            )
-        else:
-            self._update_input(input_filename)
+        self._save_results()
 
     @property
     def informer_model_exists(self):
